Update current template with description when skipping wizard
